+++ src/life4/rank_requirements/a20_plus/pearl.py
